+++ test/example/foo/bam.js
@@ -1,4 +1,0 @@
-module.exports = {
-    parentAction: 'old',
-    get: function(req, res, next) {} // jshint ignore:line
-};