+++ test/example/methods/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    get: function(req, res, next) {}, // jshint ignore:line
+    post: function(req, res, next) {}, // jshint ignore:line
+    put: function(req, res, next) {} // jshint ignore:line
+};