+++ test/example/old.js
@@ -1,4 +1,0 @@
-module.exports = {
-    pathPart: 'new',
-    get: function(req, res, next) {} // jshint ignore:line
-};