+++ test/example/orgs/repos/view.js
@@ -1,6 +1,0 @@
-module.exports = {
-    parentAction: 'view',
-    params: 'repoId',
-    pathPart: null,
-    get: function(req, res, next) {} // jshint ignore:line
-};