+++ test/example/orgs/view.js
@@ -1,5 +1,0 @@
-module.exports = {
-    params: 'orgId',
-    pathPart: null,
-    get: function(req, res, next) {} // jshint ignore:line
-};