+++ test/example/params.js
@@ -1,5 +1,0 @@
-module.exports = {
-    params: ['a', 'b'],
-    paramsAlt: 'x',
-    get: function(req, res, next) {} // jshint ignore:line
-};