+++ test/example/paramsAttribute/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    paramsAlt: 'id',
+    get: function(req, res, next) {} // jshint ignore:line
+};