+++ test/example/parentAction/foo/bar/view.js
@@ -1,0 +1,5 @@
+module.exports = {
+    parentAction: 'view',
+    params: 'barId',
+    get: function(req, res, next) {} // jshint ignore:line
+};