+++ test/example/path/edit.js
@@ -1,0 +1,4 @@
+module.exports = {
+    path: '/update',
+    get: function(req, res, next) {} // jshint ignore:line
+};