+++ test/example/path/foo/inherit.js
@@ -1,0 +1,4 @@
+module.exports = {
+    parentAction: 'edit',
+    get: function(req, res, next) {} // jshint ignore:line
+};