+++ test/example/pathPart/edit.js
@@ -1,0 +1,4 @@
+module.exports = {
+    pathPart: 'update',
+    get: function(req, res, next) {} // jshint ignore:line
+};