+++ test/example/pathPart/removed/view.js
@@ -1,0 +1,4 @@
+module.exports = {
+    params: 'id',
+    get: function(req, res, next) {} // jshint ignore:line
+};