+++ test/example/pathPart/view.js
@@ -1,0 +1,5 @@
+module.exports = {
+    params: 'id',
+    pathPart: null,
+    get: function(req, res, next) {} // jshint ignore:line
+};