+++ test/example/users/edit.js
@@ -1,5 +1,0 @@
-module.exports = {
-    params: 'id',
-    get: function(req, res, next) {}, // jshint ignore:line
-    post: function(req, res, next) {} // jshint ignore:line
-};