+++ test/example/users/new.js
@@ -1,4 +1,0 @@
-module.exports = {
-    get: function(req, res, next) {}, // jshint ignore:line
-    post: function(req, res, next) {} // jshint ignore:line
-};