+++ test/example/users/permissions/edit.js
@@ -1,5 +1,0 @@
-module.exports = {
-    parentAction: 'view',
-    params: 'permissionId',
-    get: function(req, res, next) {} // jshint ignore:line
-};