+++ test/example/users/view.js
@@ -1,5 +1,0 @@
-module.exports = {
-    pathPart: null,
-    params: 'id',
-    get: function(req, res, next) {} // jshint ignore:line
-};