+++ test/example/zoo/creatures/monkey.js
@@ -1,4 +1,0 @@
-module.exports = {
-    path: '/baboon',
-    get: function(req, res, next) {} // jshint ignore:line
-};