+++ changelog.md
@@ -26,3 +26,4 @@
 * Tests for creation of express routes
 * Refactor tests
 * README `load` options
+* README update