+++ changelog.md
@@ -17,7 +17,7 @@
 * Support definition of actions in route definition
 * Clear require cache before each test
 
-## Next
+## 1.0.0
 
 * Hooks
 * `wrapper` option